+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 * Refactor to easily [use Ed as React component](https://github.com/the-grid/ed#use).
 * `props.widgetPath` for the iframe widget base directory. Defaults to `'./node_modules/'`
+* Add `location` widget (no UI to add yet)
 
 ## 0.14.1 - 2016-06-08